pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mind-cli"
version = "0.1.0"
description = "Mind CLI - Time logging and reporting automation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Mind CLI Team" }
]
keywords = ["cli", "clockify", "jira", "time-tracking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "holidays>=0.91",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
mind = "mind.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["mind"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
